+++ redis/asyncio/client.py
@@ -253,6 +253,9 @@ class Redis(
 
         self.response_callbacks = CaseInsensitiveDict(self.__class__.RESPONSE_CALLBACKS)
 
+        if self.connection_pool.connection_kwargs.get("protocol") == "3":
+            self.response_callbacks.update(self.__class__.RESP3_RESPONSE_CALLBACKS)
+
         # If using a single connection client, we need to lock creation-of and use-of
         # the client in order to avoid race conditions such as using asyncio.gather
         # on a set of redis commands
