+++ redis/cluster.py
@@ -8,8 +8,8 @@ from typing import Any, Callable, Dict, List, Optional, Tuple, Union
 
 from redis.backoff import default_backoff
 from redis.client import CaseInsensitiveDict, PubSub, Redis, parse_scan
-from redis.commands import READ_COMMANDS, CommandsParser, RedisClusterCommands
-from redis.connection import ConnectionPool, DefaultParser, Encoder, parse_url
+from redis.commands import READ_COMMANDS, RedisClusterCommands
+from redis.connection import ConnectionPool, DefaultParser, parse_url
 from redis.crc import REDIS_CLUSTER_HASH_SLOTS, key_slot
 from redis.exceptions import (
     AskError,
@@ -29,6 +29,7 @@ from redis.exceptions import (
     TryAgainError,
 )
 from redis.lock import Lock
+from redis.parsers import CommandsParser, Encoder
 from redis.retry import Retry
 from redis.utils import (
     dict_merge,
@@ -138,6 +139,7 @@ REDIS_ALLOWED_KEYS = (
     "queue_class",
     "retry",
     "retry_on_timeout",
+    "protocol",
     "socket_connect_timeout",
     "socket_keepalive",
     "socket_keepalive_options",
