+++ redis/utils.py
@@ -13,6 +13,13 @@ except ImportError:
     HIREDIS_PACK_AVAILABLE = False
 
 try:
+    import ssl  # noqa
+
+    SSL_AVAILABLE = True
+except ImportError:
+    SSL_AVAILABLE = False
+
+try:
     import cryptography  # noqa
 
     CRYPTOGRAPHY_AVAILABLE = True
