+++ tests/test_pubsub.py
@@ -767,9 +767,9 @@ class TestBaseException:
         assert msg is not None
         # timeout waiting for another message which never arrives
         assert is_connected()
-        with patch("redis.connection.PythonParser.read_response") as mock1:
+        with patch("redis.parsers._RESP2Parser.read_response") as mock1:
             mock1.side_effect = BaseException("boom")
-            with patch("redis.connection.HiredisParser.read_response") as mock2:
+            with patch("redis.parsers._HiredisParser.read_response") as mock2:
                 mock2.side_effect = BaseException("boom")
 
                 with pytest.raises(BaseException):
